use uint32_t for quantity

+++ lib/OrderEvent.hpp
@@ -52,18 +52,18 @@ class OrderFillEvent : public OrderEventBase {
 public:
     OrderFillEvent();
     OrderFillEvent(const OrderFillEvent& event);
-    OrderFillEvent(const uint64_t eventId, const uint64_t orderId, const uint64_t timestamp, const int fillQuantity, const double fillPrice);
-    const int getFillQuantity() const { return myFillQuantity; }
+    OrderFillEvent(const uint64_t eventId, const uint64_t orderId, const uint64_t timestamp, const uint32_t fillQuantity, const double fillPrice);
+    const uint32_t getFillQuantity() const { return myFillQuantity; }
     const double getFillPrice() const { return myFillPrice; }
-    void setFillQuantity(const int fillQuantity) { myFillQuantity = fillQuantity; }
-    void setFillPrice(const int fillPrice) { myFillPrice = fillPrice; }
+    void setFillQuantity(const uint32_t fillQuantity) { myFillQuantity = fillQuantity; }
+    void setFillPrice(const double fillPrice) { myFillPrice = fillPrice; }
     virtual std::shared_ptr<OrderEventBase> clone() const override { return std::make_shared<OrderFillEvent>(*this); }
     virtual void applyTo(MarketOrder& order) const override;
     virtual void applyTo(LimitOrder& order) const override;
     virtual void init() override;
     virtual const std::string getAsJason() const override;
 private:
-    int myFillQuantity;
+    uint32_t myFillQuantity;
     double myFillPrice;
 };
 
@@ -73,7 +73,7 @@ public:
     OrderModifyPriceEvent(const OrderModifyPriceEvent& event);
     OrderModifyPriceEvent(const uint64_t eventId, const uint64_t orderId, const uint64_t timestamp, const double modifiedPrice);
     const double getModifiedPrice() const { return myModifiedPrice; }
-    void setModifiedPrice(const int modifiedPrice) { myModifiedPrice = modifiedPrice; }
+    void setModifiedPrice(const double modifiedPrice) { myModifiedPrice = modifiedPrice; }
     virtual std::shared_ptr<OrderEventBase> clone() const override { return std::make_shared<OrderModifyPriceEvent>(*this); }
     virtual void applyTo(LimitOrder& order) const override;
     virtual void init() override;
@@ -87,14 +87,14 @@ public:
     OrderModifyQuantityEvent();
     OrderModifyQuantityEvent(const OrderModifyQuantityEvent& event);
     OrderModifyQuantityEvent(const uint64_t eventId, const uint64_t orderId, const uint64_t timestamp, const double modifiedQuantity);
-    const int getModifiedQuantity() const { return myModifiedQuantity; }
-    void setModifiedQuantity(const int modifiedQuantity) { myModifiedQuantity = modifiedQuantity; }
+    const uint32_t getModifiedQuantity() const { return myModifiedQuantity; }
+    void setModifiedQuantity(const uint32_t modifiedQuantity) { myModifiedQuantity = modifiedQuantity; }
     virtual std::shared_ptr<OrderEventBase> clone() const override { return std::make_shared<OrderModifyQuantityEvent>(*this); }
     virtual void applyTo(LimitOrder& order) const override;
     virtual void init() override;
     virtual const std::string getAsJason() const override;
 private:
-    int myModifiedQuantity;
+    uint32_t myModifiedQuantity;
 };
 
 class OrderCancelEvent : public OrderEventBase {
